// core/ec-core/umbracoClient/models/BannerPropertiesModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApiLinkModel } from './ApiLinkModel';
export type BannerPropertiesModel = {
    link?: Array<ApiLinkModel> | null;
    message?: string | null;
    removable?: boolean | null;
    backgroundColor?: string | null;
};